fix: navigating to cognito signin page takes a minute (#101)
*Issue #, if available:*

*Description of changes:*

Due to CloudFront issue, it takes a minute to navigate to cognito
sign-in page. The workaround is to add some body to the redirect
response.

We cannot directly modify the redirect response because it is
encapsulated in the amplify library. That's why we use patch-package
instead.

By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice.